test: fix Oauth2 migration eror
