+++ server/integration-test/src/test/java/com/oceanbase/odc/migrate/jdbc/common/V2413ConnectPasswordMigrateTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.metadb.migrate;
+package com.oceanbase.odc.migrate.jdbc.common;
 
 import javax.sql.DataSource;
 
@@ -25,9 +25,8 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.jdbc.JdbcTestUtils;
 
 import com.oceanbase.odc.ServiceTestEnv;
-import com.oceanbase.odc.migrate.jdbc.common.V2422ConnectEmptyPasswordMigrate;
 
-public class V2422ConnectSysPasswordMigrateTest extends ServiceTestEnv {
+public class V2413ConnectPasswordMigrateTest extends ServiceTestEnv {
 
     @Autowired
     private DataSource dataSource;
@@ -48,25 +47,18 @@ public class V2422ConnectSysPasswordMigrateTest extends ServiceTestEnv {
                 + "VALUES (1, 'test1', 'test1', '123456', 'BCRYPT')");
         jdbcTemplate.update("INSERT INTO `odc_session_manager` ("
                 + "`id`, `user_id`, `session_name`, `db_mode`, `host`, `port`, `cluster`, "
-                + "`tenant`, `db_user`, `password`, `default_DBName`, `extend_info`, `salt`, `cipher`) "
-                + " VALUES (1, '1', 'test1', 'OB_MYSQL', '127.1', '2881', '', 'mysql', 'root', '', 'mysql', '{}',"
-                + " 'fakesaltvalues123', 'AES256SALT')");
+                + "`tenant`, `db_user`, `password`, `default_DBName`, `extend_info`, `salt`) "
+                + " VALUES (1, '1', 'test1', 'OB_MYSQL', '127.1', '2881', '', 'mysql', 'root', '123456', 'mysql', '{}', '')");
         jdbcTemplate.update("INSERT INTO `odc_session_extended` ("
                 + " `sid`, `session_timeout`, `sys_user`, `sys_user_password`) "
-                + " VALUES ('1', '60', 'root', '')");
+                + " VALUES ('1', '60', 'root', 'NjU0MzIx')");
 
         // execute
-        V2422ConnectEmptyPasswordMigrate migrate = new V2422ConnectEmptyPasswordMigrate();
+        V2413ConnectPasswordMigrate migrate = new V2413ConnectPasswordMigrate();
         migrate.migrate(dataSource);
 
-        // verify password
-        int count1 =
-                JdbcTestUtils.countRowsInTableWhere(jdbcTemplate, "odc_session_manager", "`password` <> ''");
-        Assert.assertEquals(1, count1);
-
-        // verify sys password
-        int count2 =
-                JdbcTestUtils.countRowsInTableWhere(jdbcTemplate, "odc_session_extended", "`sys_user_password` <> ''");
-        Assert.assertEquals(1, count2);
+        // verify
+        int count = JdbcTestUtils.countRowsInTableWhere(jdbcTemplate, "odc_session_manager", "`password` <> '123456'");
+        Assert.assertEquals(1, count);
     }
 }
